+++ src/jawl_voicecompanion/__init__.py
@@ -3,6 +3,14 @@
 from .gateway import TextGateway
 from .hostos_policy import HostOSPolicy
 from .models import AccessLevel, RiskClass, ToolRequest
+from .arbiter import TurnArbiter, TurnPriority
 
-__all__ = ["AccessLevel", "HostOSPolicy", "RiskClass", "TextGateway", "ToolRequest"]
-
+__all__ = [
+    "AccessLevel",
+    "HostOSPolicy",
+    "RiskClass",
+    "TextGateway",
+    "ToolRequest",
+    "TurnArbiter",
+    "TurnPriority",
+]